Released!

+++ acceptGJFriendRequest20.php
@@ -8,6 +8,7 @@ $targetAccountID = sqlTrim($_POST["targetAccountID"]);
 $requestID = sqlTrim($_POST["requestID"]);
 
 if(disabled($accountID)) exit("-1");
+if (!checkAct($accountID)) exit("-1");
 
 if(checkGJP($gjp, $accountID)) {
 	$q = $db->prepare("DELETE FROM frRequests WHERE requestID = :r");
