+++ deleteGJAccComment20.php
@@ -7,6 +7,7 @@ $gjp = sqlTrim($_POST["gjp"]);
 $commentID = sqlTrim($_POST["commentID"]);
 
 if(disabled($accountID)) exit("-1");
+if (!checkAct($accountID)) exit("-1");
 
 if(checkGJP($gjp, $accountID)) {
 	$q = $db->prepare("DELETE FROM comments WHERE commentID = :c AND accountID = :a");
