+++ deleteGJLevelUser20.php
@@ -7,6 +7,7 @@ $gjp = sqlTrim($_POST["gjp"]);
 $levelID = sqlTrim($_POST["levelID"]);
 
 if(disabled($accountID)) exit("-1");
+if (!checkAct($accountID)) exit("-1");
 
 if(checkGJP($gjp, $accountID)) {
 	$q = $db->prepare("DELETE FROM levels WHERE levelID = :l");
