+++ deleteGJMessages20.php
@@ -9,13 +9,14 @@ $isSender = sqlTrim($_POST["isSender"]);
 $messages = sqlTrim($_POST["messages"]);
 
 if(disabled($accountID)) exit("-1");
+if (!checkAct($accountID)) exit("-1");
 
 if(checkGJP($gjp, $accountID)) {
 	if($messageID != "") {
 		$q = $db->prepare("DELETE FROM messages WHERE messageID = :m");
 		$q->execute(array('m' => $messageID));
 		exit("1");
-	} 
+	}
 
 	if($messages != "") {
 		if($isSender != "") {
@@ -29,7 +30,7 @@ if(checkGJP($gjp, $accountID)) {
 			$mArray = explode(',', $messages);
 			foreach($mArray as $m) {
 				$q = $db->prepare("DELETE FROM messages WHERE messageID = :m");
-				$q->execute(array('m' => $messageID));
+				$q->execute(array('m' => $m));
 			}
 			exit("1");
 		}
